+++ Tests/Unit/Process/Cleaner/OldProcessCleanerTest.php
@@ -52,16 +52,42 @@ class OldProcessCleanerTest extends UnitTestCase
     }
 
     #[Test]
-    public function cleanDoesNothingWhenRepositoryReturnsNonArray(): void
+    public function cleanDoesNothingWhenRepositoryReturnsEmpty(): void
     {
+        $this->processRepository
+            ->expects(self::once())
+            ->method('getActiveProcessesOlderThanOneHour')
+            ->willReturn([]);
+
+        $this->processManager->expects(self::never())->method('processExists');
+        $this->processRepository->expects(self::never())->method('removeByProcessId');
+        $this->queueRepository->expects(self::never())->method('unsetQueueProcessId');
+
+        $this->subject->clean();
+    }
+
+    #[Test]
+    public function cleanDoesNothingWhenRepositoryReturnsNull(): void
+    {
+        $this->expectException(\UnexpectedValueException::class);
+        $this->expectExceptionMessage('Expected array, got NULL');
         $this->processRepository
             ->expects(self::once())
             ->method('getActiveProcessesOlderThanOneHour')
             ->willReturn(null);
 
-        $this->processManager->expects(self::never())->method('processExists');
-        $this->processRepository->expects(self::never())->method('removeByProcessId');
-        $this->queueRepository->expects(self::never())->method('unsetQueueProcessId');
+        $this->subject->clean();
+    }
+
+    #[Test]
+    public function ensureEarlyReturnIfResultIsNotAnArray(): void
+    {
+        $this->expectException(\UnexpectedValueException::class);
+        $this->expectExceptionMessage('Expected array, got string');
+        $this->processRepository
+            ->expects(self::once())
+            ->method('getActiveProcessesOlderThanOneHour')
+            ->willReturn('string-instead-of-array');
 
         $this->subject->clean();
     }
@@ -80,6 +106,70 @@ class OldProcessCleanerTest extends UnitTestCase
                 [
                     'system_process_id' => 1,
                     'process_id' => 'skip-me-too',
+                ],
+            ]);
+
+        $this->processManager->expects(self::never())->method('processExists');
+        $this->processRepository->expects(self::never())->method('removeByProcessId');
+        $this->queueRepository->expects(self::never())->method('unsetQueueProcessId');
+
+        $this->subject->clean();
+    }
+
+    #[Test]
+    public function cleanSkipsProcessesWithSystemIdEqualToOneButProcessTheOneAbove(): void
+    {
+        $this->processRepository
+            ->expects(self::once())
+            ->method('getActiveProcessesOlderThanOneHour')
+            ->willReturn([
+                [
+                    'system_process_id' => 1,
+                    'process_id' => 'skip-me',
+                ],
+                [
+                    'system_process_id' => 2,
+                    'process_id' => 'dont-skip-me',
+                ],
+            ]);
+
+        $this->processManager->expects(self::once())->method('processExists');
+        $this->processRepository->expects(self::once())->method('removeByProcessId');
+        $this->queueRepository->expects(self::once())->method('unsetQueueProcessId');
+
+        $this->subject->clean();
+    }
+
+    #[Test]
+    public function cleanConvertStringToInt(): void
+    {
+        $this->processRepository
+            ->expects(self::once())
+            ->method('getActiveProcessesOlderThanOneHour')
+            ->willReturn([
+                [
+                    'system_process_id' => '2',
+                    'process_id' => 'dont-skip-me',
+                ],
+            ]);
+
+        $this->processManager->expects(self::once())->method('processExists');
+        $this->processRepository->expects(self::once())->method('removeByProcessId');
+        $this->queueRepository->expects(self::once())->method('unsetQueueProcessId');
+
+        $this->subject->clean();
+    }
+
+    #[Test]
+    public function cleanConvertStringCrawlerToInt(): void
+    {
+        $this->processRepository
+            ->expects(self::once())
+            ->method('getActiveProcessesOlderThanOneHour')
+            ->willReturn([
+                [
+                    'system_process_id' => 'crawler',
+                    'process_id' => 'skip-as-i-cannot-convert-string-to-int',
                 ],
             ]);